# ARROW-10540 [Rust] Improve filtering