… rust/arrow/src/array/transform/variable_size.rs
… rust/arrow/src/compute/kernels/filter.rs