## Motivation

The current code-base to `filter` an array is typed, which means that we can't easily generalize it for arbitrary depths.

That code is also similar to the code in `take`: both these operations perform a very similar operation: from an array, produce another array taking elements from an existing array.

## This PR

This is a proposal to introduce a new `struct` (`MutableArrayData` for the lack of a better name) that is useful to efficiently construct an array taken from another array. The main use-case of this `struct` is to be used in unary operations that are type-independent, such as `filter`, `take` and `sort`.

With this `struct`, this PR:

* Adds `MutableArrayData` to efficiently create `ArrayData` as partial `memcopies` from another `ArrayData`
* Adds support to filtering of `List`s of arbitrary depth and types
* improved the benchmark to use random numbers to avoid spurious correlations (e.g. null and filter mask was 100% correlated in one of the cases). Also added benchmark for strings.
* Adds support for filtering using boolean arrays with an offset
* (only +170 LOC, even after all the new tests and benchmarks)

The gist of this `struct` can be seen in the example reproduced below:

```rust
use std::sync::Arc;
use arrow::{array::{Int32Array, Array, MutableArrayData}};

let array = Int32Array::from(vec![1, 2, 3, 4, 5]).data();

// Create a new `MutableArrayData` from an array and with a capacity.
let mut mutable = MutableArrayData::new(&array, 4);
mutable.extend(1, 3); // extend from the slice [1..3], [2,3]
mutable.extend(0, 3); // extend from the slice [0..3], [1,2,3]

// `.freeze()` to convert `MutableArrayData` into a `ArrayData`.
let new_array = Int32Array::from(Arc::new(mutable.freeze()));

assert_eq!(Int32Array::from(vec![2, 3, 1, 2, 3]), new_array);
```

## Design

There are 3 core ideas on this design:

1. it operates at the `ArrayData`, i.e. it is (rust-)type-independent and thus can be used on a variety of use-cases.
2. it assumes that it is always advantageous to memcopy a slice of size 2*N than 2 slices of size N.
3. it assumes that it is advantageous to pre-allocate memory (to the extent possible)

With this in mind, the API has 3 methods: `new`, `extend` and `freeze`. The code in `filter` was reducing memcopies to some extent, but this PR further expands on this idea by copying the largest possible chunk, instead of up to 64 slots.

## Implementation

* Replace the bit iteration in `filter.rs` by the `bitchunkIterator`
* Replace all array creation from `filter.rs` to a new module (`array/transform/`)
* Made `filter` code be focused on how to efficiently perform the filter (instead of how to build the buffers)

## Benchmarks

<details>
 <summary>Benchmarks</summary>

```bash
set -e
git checkout 010d260173a9e110901ca67372a4ac379a615a13
cargo bench --bench filter_kernels
git checkout mutable_filter
cargo bench --bench filter_kernels
```

```bash
Previous HEAD position was 010d26017 Improved filter bench.
Your branch is up to date with 'origin/mutable_filter'.
   Compiling arrow v3.0.0-SNAPSHOT (/Users/jorgecarleitao/projects/arrow/rust/arrow)
    Finished bench [optimized] target(s) in 50.14s
     Running /Users/jorgecarleitao/projects/arrow/rust/target/release/deps/filter_kernels-9f6c18b388f93061
Gnuplot not found, using plotters backend
filter u8 low selectivity                                                                            
                        time:   [489.72 us 490.02 us 490.34 us]
                        change: [+30.894% +32.676% +33.916%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe

filter u8 high selectivity                                                                             
                        time:   [6.8506 us 6.8599 us 6.8701 us]
                        change: [-57.316% -56.946% -56.621%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

filter u8 very low selectivity                                                                             
                        time:   [16.036 us 16.054 us 16.074 us]
                        change: [+102.22% +104.20% +106.38%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  5 (5.00%) high severe

filter context u8 low selectivity                                                                            
                        time:   [234.86 us 235.13 us 235.41 us]
                        change: [-35.318% -34.980% -34.651%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe

filter context u8 high selectivity                                                                             
                        time:   [4.0385 us 4.0436 us 4.0488 us]
                        change: [-61.652% -61.402% -61.115%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  3 (3.00%) high mild
  5 (5.00%) high severe

filter context u8 very low selectivity                                                                             
                        time:   [1.0964 us 1.0993 us 1.1024 us]
                        change: [-57.418% -57.154% -56.863%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

filter context u8 w NULLs low selectivity                                                                            
                        time:   [451.90 us 452.26 us 452.65 us]
                        change: [-15.196% -14.870% -14.539%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

filter context u8 w NULLs high selectivity                                                                            
                        time:   [301.93 us 304.32 us 306.65 us]
                        change: [-20.688% -20.012% -19.357%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking filter context u8 w NULLs very low selectivity: Collecting 100 samples in estimated 5.0018 s (3.7M iterations                                                                                                                          filter context u8 w NULLs very low selectivity                        
                        time:   [1.3675 us 1.3734 us 1.3793 us]
                        change: [-59.899% -59.602% -59.302%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe

filter context f32 low selectivity                                                                            
                        time:   [459.46 us 459.83 us 460.25 us]
                        change: [-16.233% -15.973% -15.697%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  9 (9.00%) high mild
  3 (3.00%) high severe

filter context f32 high selectivity                                                                            
                        time:   [303.60 us 305.43 us 307.52 us]
                        change: [-20.196% -19.416% -18.644%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

filter context f32 very low selectivity                                                                             
                        time:   [1.3427 us 1.3451 us 1.3478 us]
                        change: [-62.984% -62.749% -62.510%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe

filter context string low selectivity                                                                            
                        time:   [876.26 us 880.96 us 887.54 us]
                        change: [-15.814% -15.375% -14.900%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) high mild
  6 (6.00%) high severe

Benchmarking filter context string high selectivity: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.1s, enable flat sampling, or reduce sample count to 50.
filter context string high selectivity                                                                             
                        time:   [1.4070 ms 1.4094 ms 1.4120 ms]
                        change: [-15.332% -14.829% -14.364%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

filter context string very low selectivity                                                                             
                        time:   [7.2640 us 7.2763 us 7.2888 us]
                        change: [+73.534% +75.057% +76.543%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe
```
</details>