@jorgecarleitao 2020-11-14
I have rebased against latest master and further improved the code. IMO this is now ready to review: benchmarks are now better for all cases where we pre-compute the filter except for 1 case (very low selectivity) where it is 1.5x worse. 

For a single filter, I know how to improve it: we need to not pre-compute the filter and instead use an iterator.

I can continue to investigate this, but IMO this is an improvement to the code base and performance.