@jorgecarleitao 2020-11-15
FYI I was unable to make this sufficiently efficient for the `take`: it has a large performance hit when compared to the current approach. As such, I am not sure if this can be used more generally. This may be moved to the filter kernel and remain there, and leave the `take` as a separate implementation.