@jorgecarleitao 2020-11-17
This code will be re-used to implement `merge-sort` and `join`. Some notes about this here: https://docs.google.com/document/d/1KKuBvfx7uKi-x-tWOL60R1FNjDW8B790zPAv6yAlYcU/edit?usp=sharing