@alamb 2020-11-19
FWIW keeping "filter" as a special case that is very fast is not unreasonable -- it is likely to be one of the most performance critical pieces of code in analytics systems, so it is naturally likely to end up as the most complex / optimized.